requirements.txt
absl-py==2.2.2
accelerate==1.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.8
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
datasets==2.20.0
debugpy==1.8.17
decorator==5.2.1
dictdot==1.5.1
diffusers==0.32.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.1
executing==2.2.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.60.0
frozenlist==1.5.0
fsspec==2024.5.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.27.1
idna==3.4
importlib_metadata==8.6.1
iopath==0.1.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.9.2
kiwisolver==1.4.9
libclang==18.1.1
liger_kernel==0.6.2
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
omegaconf==2.3.0
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
portalocker==3.2.0
prompt_toolkit==3.0.52
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyarrow-hotfix==0.6
Pygments==2.19.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.15.2
sentry-sdk==2.26.1
setproctitle==1.3.5
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
timm==1.0.12
tokenizers==0.21.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
triton==3.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.13
wandb==0.17.6
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xformer==1.0.1
xformers==0.0.32.post2
xxhash==3.5.0
yacs==0.1.8
yarl==1.20.0
zipp==3.21.0
nvidia-cuda-nvcc-cu12==12.8.93